# course_selection/openapi.yaml
openapi: 3.0.3
info: {title: Academic Selection Workbench, version: 0.1.0}
paths:
  /api/state:
    get:
      responses:
        '200':
          description: Local persisted workbench state
          content: {application/json: {schema: {$ref: '#/components/schemas/WorkbenchState'}}}
  /api/login-configuration:
    post:
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginConfigurationRequest'}}}}
      responses: {'201': {description: DPAPI-protected login saved}, '400': {description: Invalid login configuration}}
    delete:
      responses: {'204': {description: Saved login cleared}}
  /api/tasks:
    post:
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TaskRequest'}}}}
      responses: {'202': {description: Accepted}}
  /api/plans:
    post:
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PlanRequest'}}}}
      responses: {'201': {description: Persisted read-only plan}}
  /api/plans/latest:
    get:
      responses: {'200': {description: Latest persisted read-only plan}, '404': {description: No plan}}
  /api/executions/selection:
    post:
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SelectionExecutionRequest'}}}}
      responses: {'202': {description: Explicitly confirmed single execution queued}, '409': {description: Confirmation or applicability check failed}}
  /api/tasks/{identity}:
    get:
      parameters: [{name: identity, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Task state}, '404': {description: Task not found}}
    delete:
      parameters: [{name: identity, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Cancellation requested}, '409': {description: Task cannot be cancelled}}
  /api/tasks/{identity}/finish:
    post:
      parameters: [{name: identity, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Observation finish requested}, '409': {description: Task cannot be finished}}
  /api/notices/candidates:
    get:
      responses: {'200': {description: Candidate notice versions}}
    post:
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
      responses: {'201': {description: Candidate notice saved}}
  /api/notices/discover:
    post:
      requestBody: {required: false, content: {application/json: {schema: {type: object, properties: {index_url: {type: string, format: uri}}}}}}
      responses: {'201': {description: Matching official notice pages downloaded, parsed, and saved}, '400': {description: Official notice source could not be read}}
  /api/notices/{identity}/confirm:
    post:
      parameters: [{name: identity, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Notice confirmed}, '409': {description: Notice cannot be confirmed}}
  /api/timetable/import:
    post:
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, properties: {timetable: {type: string, format: binary}, term: {type: string}}}}}}
      responses: {'201': {description: Timetable snapshot imported}}
components:
  schemas:
    LoginConfigurationRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, maxLength: 64}
        password: {type: string, format: password, maxLength: 256}
    TaskRequest:
      type: object
      required: [operation]
      properties:
        operation: {type: string, enum: [connect, refresh-selection, refresh-timetable, refresh-progress, observe-navigation]}
        context: {type: object, additionalProperties: true}
    PlanRequest:
      type: object
      required: [goals]
      properties:
        goals: {type: array, items: {type: object, additionalProperties: true}}
    SelectionExecutionRequest:
      type: object
      required: [section_id, snapshot_id, confirmation]
      properties:
        section_id: {type: string}
        snapshot_id: {type: string}
        confirmation: {type: string, description: Must exactly equal section_id}
    LoginConfiguration:
      type: object
      required: [state, configured]
      properties:
        state: {type: string, enum: [missing, configured, invalid]}
        configured: {type: boolean}
        masked_username: {type: string}
        message: {type: string}
    Task:
      type: object
      required: [id, state]
      properties:
        id: {type: string}
        operation: {type: string}
        task_kind: {type: string, enum: [observation, execution]}
        state: {type: string}
        progress: {type: object, additionalProperties: true}
        error: {type: string}
    CandidateNotice:
      type: object
      required: [version_id, status]
      properties:
        version_id: {type: string}
        status: {type: string}
        title: {type: string}
        term: {type: string}
        query_eligible: {type: boolean}
      additionalProperties: true
    Snapshot:
      type: object
      required: [id, kind, term, source, source_at, payload]
      properties:
        id: {type: string}
        kind: {type: string}
        term: {type: string}
        profile_id: {type: string, nullable: true}
        notice_id: {type: string, nullable: true}
        source: {type: string}
        source_at: {type: string}
        created_at: {type: string}
        payload: {type: object, additionalProperties: true}
    CompletedCourseFact:
      type: object
      required: [code, name, category, credits]
      properties:
        code: {type: string}
        name: {type: string}
        category: {type: string}
        credits: {type: number}
    RequirementProgress:
      type: object
      required: [key, label, required_credits, completed_credits, remaining_credits, courses]
      properties:
        key: {type: string}
        label: {type: string}
        required_credits: {type: number}
        completed_credits: {type: number}
        remaining_credits: {type: number}
        courses: {type: array, items: {$ref: '#/components/schemas/CompletedCourseFact'}}
    GraduationProgressReport:
      type: object
      required: [data_complete, progress]
      properties:
        generated_at: {type: string}
        baseline_version: {type: string}
        data_complete: {type: boolean}
        progress: {type: array, items: {$ref: '#/components/schemas/RequirementProgress'}}
    GraduationProgress:
      type: object
      required: [status, report]
      properties:
        status: {type: string, enum: [ready, incomplete, missing, invalid, not_applicable]}
        report: {$ref: '#/components/schemas/GraduationProgressReport', nullable: true}
        snapshot: {$ref: '#/components/schemas/Snapshot', nullable: true}
    WorkbenchState:
      type: object
      required: [login_configuration, profile, confirmed_notice, snapshots, snapshot_changes, latest_plan, graduation_progress, stale, academic_session, csrf_token]
      properties:
        login_configuration: {$ref: '#/components/schemas/LoginConfiguration'}
        profile: {type: object, nullable: true, additionalProperties: true}
        confirmed_notice: {type: object, nullable: true, additionalProperties: true}
        snapshots:
          type: object
          required: [selection, timetable, progress]
          properties:
            selection: {$ref: '#/components/schemas/Snapshot', nullable: true}
            timetable: {$ref: '#/components/schemas/Snapshot', nullable: true}
            progress: {$ref: '#/components/schemas/Snapshot', nullable: true}
        snapshot_changes:
          type: object
          required: [selection, timetable, progress]
          properties:
            selection: {type: object, nullable: true, additionalProperties: true}
            timetable: {type: object, nullable: true, additionalProperties: true}
            progress: {type: object, nullable: true, additionalProperties: true}
        latest_plan: {type: object, nullable: true, additionalProperties: true}
        graduation_progress: {$ref: '#/components/schemas/GraduationProgress'}
        stale:
          type: object
          required: [selection, timetable, progress]
          properties:
            selection: {type: boolean}
            timetable: {type: boolean}
            progress: {type: boolean}
        academic_session:
          type: object
          required: [state]
          properties: {state: {type: string}}
        csrf_token: {type: string}